make more types private
